Changed the code as recommended
