+++ backend/games/finance/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from views import GetProductsLibraryRetrieve
+
+
+urlpatterns = [
+    path("finance/library/<user>", GetProductsLibraryRetrieve.as_view(),
+         name="all_products_in_library")
+]
