+++ backend/games/finance/views.py
@@ -1,3 +1,11 @@
 from django.shortcuts import render
+from rest_framework import generics
+from finance.models.library import Library
+from finance.serializers import LibrarySerializer
 
-# Create your views here.
+
+class GetProductsLibraryRetrieve(generics.RetrieveAPIView):
+
+    serializer_class = LibrarySerializer
+    queryset = Library.objects.all()
+    lookup_field = 'user'
