+++ backend/games/utils/urls.py
@@ -3,8 +3,7 @@ from django.urls import path
 from utils.views import (
     GetOperatingSystemListView,
     GetGenreListView,
-    GetMinMaxPriceListView,
-    GetProductsLibraryRetrieve
+    GetMinMaxPriceListView
 )
 
 
@@ -14,6 +13,4 @@ urlpatterns = [
         "filters/platforms", GetOperatingSystemListView.as_view(), name="all_platforms"
     ),
     path("filters/prices", GetMinMaxPriceListView.as_view(), name="min_max_price"),
-    path("filters/library/<user>", GetProductsLibraryRetrieve.as_view(),
-         name="all_products_in_library")
 ]
