+++ backend/games/utils/views.py
@@ -9,9 +9,6 @@ from core.serializers import OperatingSystemSerializer
 from reference.models.genres import Genre, SubGenre
 from reference.serializers import GenreGamesSerializer
 
-from finance.models.library import Library
-from finance.serializers import LibrarySerializer
-
 from .serializers import MinMaxPriceSerializer
 
 
@@ -73,10 +70,3 @@ class GetMinMaxPriceListView(generics.GenericAPIView):
         serializer = self.get_serializer(data={})
         serializer.is_valid(raise_exception=True)
         return Response(serializer.data)
-
-
-class GetProductsLibraryRetrieve(generics.RetrieveAPIView):
-
-    serializer_class = LibrarySerializer
-    queryset = Library.objects.all()
-    lookup_field = 'user'
